Catch potential exception when GoogleApiClient is not connected but cast session exists
Cast session manager is built on top of GoogleApiClient and assumes api client
to be connected when sendMessage function is called. We've discovered that in some cases
apiClient might be disconnected which causes IllegalStateException. Since session manager
doesn't provide any public interface for checking client state, the only thing we can do -
wrap the function into try-catch block and give time for the client to reconnect before next call

An alternative solution would be using GoogleApiClient directly,
but it is not recommended since we have Cast Application Framework